Fix CSP issue (for things like Firefox OS)
No one seems to know what the best way to do get the global object is
but it seems that just "this" should work fine until someone can unravel
the mess that is JavaScript.  As far as I know it will only fail if
someone adds "use strict" to the top of the .js file.